cleanup: hides the browsersync notification.
Having it show up in the iframe was distracting, and I couldn't find a
way to conditionally hide it just in the iframe.

Another approach would be to disable browsersync entirely for pages in
the examples folder using
https://www.browsersync.io/docs/options/#option-snippetOptions. That
would make sense in some ways, since those files aren't being compiled
or watched, but it might still be nice for things like cross-browser
testing. I'll leave it there for now.